Add pages files
